image transform for cdn

+++ dags/musinsa/ops/transform_images.py
@@ -2,6 +2,7 @@ import boto3
 import logging
 import os
 
+from botocore.exceptions import ClientError
 from airflow.models import BaseOperator
 from airflow.utils.decorators import apply_defaults
 from airflow.providers.amazon.aws.hooks.s3 import S3Hook 
@@ -11,7 +12,7 @@ import urllib.request
 
 logger = logging.getLogger(__name__)
 
-class ImageUploadOperator(BaseOperator):
+class TransformImageOperator(BaseOperator):
         
     @apply_defaults
     def __init__(self, *args, **kwargs):
@@ -35,24 +36,34 @@ class ImageUploadOperator(BaseOperator):
         url_dict = {}
         
         for key in xcomData:
-            
             url_list = []
+            file_name = "tmp.jpg"
             
             for i, url in enumerate(xcomData[key]):
-                file_name = "tmp.jpg"
                 file_key = f"musinsa/{key}/{i}.webp"
-                urllib.request.urlretrieve(url, file_name)
-
-                self.s3_client.upload_file(
-                    Filename=file_name,
-                    Bucket=self.bucket_name,
-                    Key=file_key,
-                    ExtraArgs=self.extra_args
-                )
-                url = f"https://{self.bucket_name}.s3.ap-northeast-2.amazonaws.com/{file_key}"
-                url_list.append(url)
                 
-                os.remove(file_name)
+                # Check if the file already exists in S3
+                try:
+                    self.s3_client.head_object(Bucket=self.bucket_name, Key=file_key)
+                    logger.info(f"File already exists: {file_key}. Skipping upload.")
+                    url = f"https://{self.bucket_name}.s3.ap-northeast-2.amazonaws.com/{file_key}"
+                    url_list.append(url)
+                    continue
+                except ClientError as e:
+                    # The error signifies that the object does not exist
+                    if e.response['Error']['Code'] == "404":
+                        logger.info(f"Uploading file: {file_key}")
+                        urllib.request.urlretrieve(url, file_name)
+                        self.s3_client.upload_file(
+                            Filename=file_name,
+                            Bucket=self.bucket_name,
+                            Key=file_key,
+                            ExtraArgs=self.extra_args
+                        )
+                        url = f"https://{self.bucket_name}.s3.ap-northeast-2.amazonaws.com/{file_key}"
+                        url_list.append(url)
+                        
+                        os.remove(file_name)
                 
             url_dict[key] = url_list
